requirements.txt
annotated-types==0.5.0
anyio==3.7.1
Authlib==1.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.100.0
Flask==2.3.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
nltk==3.8.1
oauthlib==3.2.2
protobuf==3.20.3
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
pymongo==4.4.0
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.2
uvicorn==0.22.0
Werkzeug==2.3.4
zipp==3.15.0
